# src/main/resources/static/swagger-apis/api1/swagger.yaml
swagger: "2.0"
info:
  description: "Paypal Rest API Service"
  version: "1.0.0"
  title: "Paypal Rest API Service for payment transactions."
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "sandeeprgurav@gmail.com"
basePath: ""
tags:
  - name: "invoice"
    description: "Invoice Information"
  - name: "billing"
    description: "Billing Information"
  - name: "subscribe"
    description: "Subscribe Information"    
schemes:
  - "https"
  - "http"
paths:
  /invoice/all:
    get:
      tags:
        - "invoice"
      summary: "List Invoices by Process Date"
      description: "Invoice API to find all invoices for provided process date."
      consumes:
        - "application/json"
        - "application/xml"
      produces:
        - "application/xml"
        - "application/json"
      parameters:
      - name: "processDate"
        in: "query"
        description: "Process Date"
        required: true
        type: "string"        
  /invoice/filterByCustomerIdAndProcessDate:
    get:
      tags:
        - "invoice"
      summary: "Filter by CustomerId and Process Date "
      description: "Filter by Invoices for provided customerId and Process Date."
      consumes:
        - "application/json"
        - "application/xml"
      produces:
        - "application/xml"
        - "application/json"
      parameters:
      - name: "customerId"
        in: "query"
        description: "Customer Id"
        required: true
        type: "string"
      - name: "processDate"
        in: "query"
        description: "Process Date"
        required: true
        type: "string" 
  /invoice/filterByCustomerIdAndProcessYearMonth:
    get:
      tags:
        - "invoice"
      summary: "Filter by CustomerId and Process YearMonth"
      description: "Filter by Invoices for provided customerId and Process YearMonth."
      consumes:
        - "application/json"
        - "application/xml"
      produces:
        - "application/xml"
        - "application/json"
      parameters:
      - name: "customerId"
        in: "query"
        description: "Customer Id"
        required: true
        type: "string"
      - name: "processYearMonth"
        in: "query"
        description: "Process YearMonth"
        required: true
        type: "string" 
  /invoice/filterByCustomerIdAndProcessYear:
    get:
      tags:
        - "invoice"
      summary: "Filter by CustomerId and Process Year"
      description: "Filter Invoices for provided customerId and Process Year."
      consumes:
        - "application/json"
        - "application/xml"
      produces:
        - "application/xml"
        - "application/json"
      parameters:
      - name: "customerId"
        in: "query"
        description: "Customer Id"
        required: true
        type: "string"
      - name: "processYear"
        in: "query"
        description: "Process Year"
        required: true
        type: "string"                        
  /billing/all:
    get:
      tags:
        - "billing"
      summary: "Billing API"
      description: "Billing API"
      consumes:
        - "application/json"
        - "application/xml"
      produces:
        - "application/xml"
        - "application/json"
      parameters:
      - name: "processDate"
        in: "query"
        description: "Process Date"
        required: true
        type: "string"        
  /subscribe/all:
    get:
      tags:
        - "subscribe"
      summary: "Subscribe API"
      description: "Subscribe API"
      consumes:
        - "application/json"
        - "application/xml"
      produces:
        - "application/xml"
        - "application/json"
      parameters:
      - name: "processDate"
        in: "query"
        description: "Process Date"
        required: true
        type: "string"        
        